Screenshots work!
I am replacing the window's contentView with the image, which is definitely not right. When I do this, it breaks dragging (and possibly resizing?), but if I append the image as a childview of the contentview, the image never shows up and idk why. but close! there's also a 'clear hints' button in the popover.