Watch for screen updates, make key shortcut consistent
To make sure things continue to work as a user resizes, adds, and removes screens, we need to watch for didChangeScreenParametersNotification (at which point we redraw all of our secret windows).
I also wanted to make sure that the global shortcut for taking a hint is the same one you can use when the menu opens up, and it seems that in order to do that, you have to disable the global one while the menu is open.